CNF_g* Reshuffle
